chore: font and SEO layout fixes
