--- pages/index.js
@@ -1,13 +1,0 @@
-const Home = () => (
-  <div className="p-4 shadow rounded bg-white">
-    <h1 className="text-primary">Scizzor</h1>
-    <h2 className="text-secondary">Scizzor</h2>
-    <h3 className="text-tertiary">Scizzor</h3>
-    <h4 className="text-alt-1">Scizzor</h4>
-    <p className="p-lg text-alt-2">Scizzor</p>
-    <p>Scizzor</p>
-    <p className="p-sm">Scizzor</p>
-  </div>
-);
-
-export default Home;
